diplomat_core

Module ast

Re-exports

pub use attrs::Attrs;

Modules

attrs

As part of the macro expansion and code generation process, Diplomat generates a simplified version of the Rust AST that captures special types such as opaque structs, Box, and Result with utilities for handling such types. This module contains utilities for dealing with Rust attributes

docs 🔒

enums 🔒

functions 🔒

idents 🔒

lifetimes 🔒

macros 🔒

methods 🔒

modules 🔒

opaque 🔒

paths 🔒

structs 🔒

traits 🔒

types 🔒

Structs

BorrowedParams

Parameters in a method that might be borrowed in the return type.

Docs

DocsUrlGenerator

Enum

A fieldless enum declaration in an FFI module.

File

Function

Represents a free function not associated with any types.

Ident

An identifier, analogous to syn::Ident and proc_macro2::Ident.

LifetimeEnv

A lifetime dependency graph used for tracking which lifetimes outlive, and are outlived by, other lifetimes.

LifetimeTransitivity

Collect all lifetimes that are either longer_or_shorter

MacroDef

Struct for defining a macro (i.e., macro_rules! example)

MacroUse

A site where a macro is used (i.e., example!(...)) Constructed in MacroUse::parse (called by MacroDef::evaluate) when a previously defined macro is used. We use the definition of MacroDef to then construct the args in the used macro.

Macros

Method

A method declared in the impl associated with an FFI struct. Includes both static and non-static methods, which can be distinguished by inspecting Method::self_param.

Module

NamedLifetime

A named lifetime, e.g. 'a.

OpaqueType

A type annotated with [diplomat::opaque] whose fields/variants are not visible. Opaque types cannot be passed by-value across the FFI boundary, so they must be boxed or passed as references.

Param

A parameter taken by a Method, not including self.

Path

PathType

A named type that is just a path, e.g. std::borrow::Cow<'a, T>.

RustLink

SelfParam

The self parameter taken by a Method.

Struct

A struct declaration in an FFI module that is not opaque.

Trait

A trait declaration in an FFI module.

TraitMethod

TraitSelfParam

The self parameter taken by a [TraitMethod].

Enums

CustomType

A type declared inside a Diplomat-annotated module.

DocType

DocsTypeReferenceSyntax

Lifetime

A lifetime, analogous to [syn::Lifetime].

LifetimeOrigin

ModSymbol

A symbol declared in a module, which can either be a pointer to another path, or a custom type defined directly inside that module

Mutability

PrimitiveType

A built-in Rust primitive scalar type.

RustLinkDisplay

StdlibOrDiplomat

For types like Result/DiplomatResult, &[T]/DiplomatSlice<T> which can be specified using (non-ffi-safe) Rust stdlib types, or FFI-safe repr(C) types from diplomat_runtime, this tracks which of the two were used.

StringEncoding

TypeName

A local type reference, such as the type of a field, parameter, or return value. Unlike CustomType, which represents a type declaration, TypeNames can compose types through references and boxing, and can also capture unresolved paths.