diplomat_core::ast::attrs

Struct Attrs

#[non_exhaustive]
pub struct Attrs {
    pub cfg: Vec<Attribute>,
    pub attrs: Vec<DiplomatBackendAttr>,
    pub skip_if_ast: bool,
    pub abi_rename: RenameAttr,
}

The list of attributes on a type. All attributes except attrs (HIR attrs) are potentially read by the diplomat macro and the AST backends, anything that is not should be added as an HIR attribute ([crate::hir::Attrs]).

Inheritance

Attributes are typically โinheritedโ: the attributes on a module apply to all types and methods with it, the attributes on an impl apply to all methods in it, and the attributes on an enum apply to all variants within it. This allows the user to specify a single attribute to affect multiple fields.

However, the details of inheritance are not always the same for each attribute. For example, rename attributes on a module only apply to the types within it (others methods would get doubly renamed).

Each attribute here documents its inheritance behavior. Note that the HIR attributes do not get inherited during AST construction, since at that time itโs unclear which of those attributes are actually available.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

cfg: Vec<Attribute>

The regular #[cfg()] attributes. Inherited, though the inheritance onto methods is the only relevant one here.

attrs: Vec<DiplomatBackendAttr>

HIR backend attributes.

Inherited, but only during lowering. See [crate::hir::Attrs] for details on which HIR attributes are inherited.

During AST attribute inheritance, HIR backend attributes are copied over from impls to their methods since the HIR does not see the impl blocks.

skip_if_ast: bool

AST backends only. For using features that may panic AST backends, like returning references.

This isnโt a regular attribute since AST backends do not handle regular attributes. Do not use in HIR backends,

Not inherited

abi_rename: RenameAttr

Renames to apply to the underlying C symbol. Can be found on methods, impls, and bridge modules, and is inherited.

Affects method names when inherited onto methods.

Affects destructor names when inherited onto types.

Inherited.

Implementations

impl Attrs

fn add_attr(&mut self, attr: Attr)

pub(crate) fn attrs_for_inheritance(&self, context: AttrInheritContext) -> Self

Get a copy of these attributes for use in inheritance, masking out things that should not be inherited

pub(crate) fn add_attrs(&mut self, attrs: &[Attribute])

pub(crate) fn from_attrs(attrs: &[Attribute]) -> Self