Enum diplomat_core::ast::Lifetime

#[non_exhaustive]
pub enum Lifetime {
    Static,
    Named(NamedLifetime),
    Anonymous,
}

A lifetime, analogous to [syn::Lifetime].

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Static

The 'static lifetime.

Named(NamedLifetime)

A named lifetime, like 'a.

Anonymous

An elided lifetime.

Implementations

impl Lifetime

pub fn to_named(self) -> Option<NamedLifetime>

Returns the inner NamedLifetime if the lifetime is the Named variant, otherwise None.

pub fn as_named(&self) -> Option<&NamedLifetime>

Returns a reference to the inner NamedLifetime if the lifetime is the Named variant, otherwise None.

impl Lifetime

pub fn to_syn(&self) -> Option<Lifetime>

Converts the Lifetime back into an AST node that can be spliced into a program.