diplomat_core::ast

Enum TypeName

#[non_exhaustive]
pub enum TypeName {

    Primitive(PrimitiveType),
    Named(PathType),
    Reference(Lifetime, Mutability, Box<TypeName>),
    Box(Box<TypeName>),
    Option(Box<TypeName>, StdlibOrDiplomat),
    Result(Box<TypeName>, Box<TypeName>, StdlibOrDiplomat),
    Write,
    StrReference(Option<Lifetime>, StringEncoding, StdlibOrDiplomat),
    PrimitiveSlice(Option<(Lifetime, Mutability)>, PrimitiveType, StdlibOrDiplomat),
    StrSlice(StringEncoding, StdlibOrDiplomat),
    CustomTypeSlice(Option<(Lifetime, Mutability)>, Box<TypeName>),
    Unit,
    SelfType(PathType),
    Ordering,
    Function(Vec<Box<TypeName>>, Box<TypeName>, Mutability),
    ImplTrait(PathType),

}

A local type reference, such as the type of a field, parameter, or return value. Unlike CustomType, which represents a type declaration, TypeNames can compose types through references and boxing, and can also capture unresolved paths.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Primitive(PrimitiveType)

A built-in Rust scalar primitive.

Named(PathType)

An unresolved path to a custom type, which can be resolved after all types are collected with [TypeName::resolve()].

Reference(Lifetime, Mutability, Box<TypeName>)

An optionally mutable reference to another type.

Box(Box<TypeName>)

A Box<T> type.

Option(Box<TypeName>, StdlibOrDiplomat)

An Option<T> or DiplomatOption type.

Result(Box<TypeName>, Box<TypeName>, StdlibOrDiplomat)

A Result<T, E> or diplomat_runtime::DiplomatResult type.

Write

StrReference(Option<Lifetime>, StringEncoding, StdlibOrDiplomat)

A &DiplomatStr or Box<DiplomatStr> type. Owned strings don’t have a lifetime.

If StdlibOrDiplomat::Stdlib, it’s specified using Rust pointer types (&T, Box), if StdlibOrDiplomat::Diplomat, it’s specified using DiplomatStrSlice, etc

PrimitiveSlice(Option<(Lifetime, Mutability)>, PrimitiveType, StdlibOrDiplomat)

A &[T] or Box<[T]> type, where T is a primitive. Owned slices don’t have a lifetime or mutability.

If StdlibOrDiplomat::Stdlib, it’s specified using Rust pointer types (&T, Box), if StdlibOrDiplomat::Diplomat, it’s specified using DiplomatSlice/DiplomatOwnedSlice/DiplomatSliceMut

StrSlice(StringEncoding, StdlibOrDiplomat)

&[DiplomatStrSlice], etc. Equivalent to &[&str]

If StdlibOrDiplomat::Stdlib, it’s specified as &[&DiplomatFoo], if StdlibOrDiplomat::Diplomat it’s specified as DiplomatSlice<&DiplomatFoo>

CustomTypeSlice(Option<(Lifetime, Mutability)>, Box<TypeName>)

&[Struct]. Meant for passing slices of structs where the struct’s layout is known to be shared between Rust and the backend language. Primarily meant for large lists of compound types like Vector3f64 or Color4i16. This is implemented on a per-backend basis.

Unit

The () type.

SelfType(PathType)

The Self type.

Ordering

std::cmp::Ordering or core::cmp::Ordering

The path must be present! Ordering will be parsed as an AST type!

Function(Vec<Box<TypeName>>, Box<TypeName>, Mutability)

ImplTrait(PathType)

Implementations

impl TypeName

pub fn is_ffi_safe(&self) -> bool

Is this type safe to be passed across the FFI boundary?

This also marks DiplomatOption<&T> as FFI-unsafe: these are technically safe from an ABI standpoint however Diplomat always expects these to be equivalent to a nullable pointer, so Option<&T> is required.

pub fn ffi_safe_version(&self) -> TypeName

What’s the FFI safe version of this type?

This also marks DiplomatOption<&T> as FFI-unsafe: these are technically safe from an ABI standpoint however Diplomat always expects these to be equivalent to a nullable pointer, so Option<&T> is required.

pub fn to_syn(&self) -> Type

Converts the TypeName back into an AST node that can be spliced into a program.

pub fn from_syn(ty: &Type, self_path_type: Option<PathType>) -> TypeName

Extract a TypeName from a syn::Type AST node. The following rules are used to infer TypeName variants:

If the type is a path with a single element that is the name of a Rust primitive, returns a TypeName::Primitive

If the type is a path with a single element Box, returns a TypeName::Box with the type parameter recursively converted

If the type is a path with a single element Option, returns a TypeName::Option with the type parameter recursively converted

If the type is a path with a single element Self and self_path_type is provided, returns a TypeName::Named

If the type is a path with a single element Result, returns a TypeName::Result with the type parameters recursively converted

If the type is a path equal to [diplomat_runtime::DiplomatResult], returns a [TypeName::DiplomatResult] with the type parameters recursively converted

If the type is a path equal to [diplomat_runtime::DiplomatWrite], returns a TypeName::Write

If the type is a owned or borrowed string type, returns a TypeName::StrReference

If the type is a owned or borrowed slice of a Rust primitive, returns a TypeName::PrimitiveSlice

If the type is a reference (& or &mut), returns a TypeName::Reference with the referenced type recursively converted

Otherwise, assume that the reference is to a CustomType in either the current module or another one, returns a TypeName::Named

pub fn is_self(&self) -> bool

Returns true if self is the TypeName::SelfType variant, otherwise false.

pub fn visit_lifetimes<'a, F, B>(&'a self, visit: &mut F) -> ControlFlow<B>
where F: FnMut(&'a Lifetime, LifetimeOrigin) -> ControlFlow<B>,

Recurse down the type tree, visiting all lifetimes.

Using this function, you can collect all the lifetimes into a collection, or examine each one without having to make any additional allocations.

pub fn any_lifetime<'a, F>(&'a self, f: F) -> bool
where F: FnMut(&'a Lifetime, LifetimeOrigin) -> bool,

Returns true if any lifetime satisfies a predicate, otherwise false.

This method is short-circuiting, meaning that if the predicate ever succeeds, it will return immediately.

pub fn all_lifetimes<'a, F>(&'a self, f: F) -> bool
where F: FnMut(&'a Lifetime, LifetimeOrigin) -> bool,

Returns true if all lifetimes satisfy a predicate, otherwise false.

This method is short-circuiting, meaning that if the predicate ever fails, it will return immediately.

pub fn longer_lifetimes<'env>( &self, lifetime_env: &'env LifetimeEnv, ) -> Vec<&'env NamedLifetime>

Returns all lifetimes in a LifetimeEnv that must live at least as long as the type.

pub fn shorter_lifetimes<'env>( &self, lifetime_env: &'env LifetimeEnv, ) -> Vec<&'env NamedLifetime>

Returns all lifetimes in a LifetimeEnv that are outlived by the type.

fn transitive_lifetime_bounds<'env>( &self, transitivity: LifetimeTransitivity<'env>, ) -> Vec<&'env NamedLifetime>

Visits the provided LifetimeTransitivity value with all NamedLifetimes in the type tree, and returns the transitively reachable lifetimes.

pub fn is_zst(&self) -> bool

pub fn is_pointer(&self) -> bool