diplomat_core::ast

Module lifetimes

Structs

LifetimeEnv

A lifetime dependency graph used for tracking which lifetimes outlive, and are outlived by, other lifetimes.

LifetimeNode 🔒

A lifetime, along with ptrs to all lifetimes that are explicitly shorter/longer than it.

LifetimeTransitivity

Collect all lifetimes that are either longer_or_shorter

NamedLifetime

A named lifetime, e.g. 'a.

Enums

Lifetime

A lifetime, analogous to [syn::Lifetime].

LongerOrShorter 🔒

A helper type for LifetimeTransitivity determining whether to find the transitively longer or transitively shorter lifetimes.