diplomat_core::ast::lifetimes

Struct LifetimeEnv

pub struct LifetimeEnv {
    pub(crate) nodes: Vec<LifetimeNode>,
}

A lifetime dependency graph used for tracking which lifetimes outlive, and are outlived by, other lifetimes.

It is similar to [syn::LifetimeDef], except it can also track lifetime bounds defined in the where clause.

Fields
nodes: Vec<LifetimeNode>
Implementations

impl LifetimeEnv

fn new() -> Self

Construct an empty LifetimeEnv.

To create one outside of this module, use LifetimeEnv::from_method_item or LifetimeEnv::from on &syn::Generics.

pub fn names(&self) -> impl Iterator<Item = &NamedLifetime> + Clone

Iterate through the names of the lifetimes in scope.

pub fn from_method_item( method: &ImplItemFn, impl_generics: Option<&Generics>, self_param: Option<&SelfParam>, params: &[Param], return_type: Option<&TypeName>, ) -> Self

Returns a LifetimeEnv for a method, accounting for lifetimes and bounds defined in both the impl block and the method, as well as implicit lifetime bounds in the optional self param, other param, and optional return type. For example, the type &'a Foo<'b> implies 'b: 'a.

pub fn from_struct_item( strct: &ItemStruct, fields: &[(Ident, TypeName, Docs)], ) -> Self

Returns a LifetimeEnv for a struct, accounding for lifetimes and bounds defined in the struct generics, as well as implicit lifetime bounds in the structโs fields. For example, the field &'a Foo<'b> implies 'b: 'a.

fn extend_implicit_lifetime_bounds( &mut self, typ: &TypeName, behind_ref: Option<&NamedLifetime>, )

Traverse a type, adding any implicit lifetime bounds that arise from having a reference to an opaque containing a lifetime. For example, the type &'a Foo<'b> implies 'b: 'a.

fn extend_generics(&mut self, generics: &Generics)

Add the lifetimes from generic parameters and where bounds.

pub fn len(&self) -> usize

Returns the number of lifetimes in the graph.

pub fn is_empty(&self) -> bool

Returns true if the graph contains no lifetimes.

pub fn lifetimes_to_tokens(&self) -> TokenStream

<'a, 'b, 'c>

Write the existing lifetimes, excluding bounds, as generic parameters.

To include lifetime bounds, use [LifetimeEnv::lifetime_defs_to_tokens].

pub(crate) fn id<L>(&self, lifetime: &L) -> Option<usize>
where NamedLifetime: PartialEq<L>,

Returns the index of a lifetime in the graph, or None if the lifetime isnโt in the graph.

fn extend_lifetimes<'a, L, I>(&mut self, iter: I)
where NamedLifetime: PartialEq<L> + From<&'a L>, L: 'a, I: IntoIterator<Item = &'a L>,

Add isolated lifetimes to the graph.

fn extend_bounds<'a, L, B, I>(&mut self, iter: I)
where NamedLifetime: PartialEq<L> + From<&'a L>, L: 'a, B: IntoIterator<Item = &'a L>, I: IntoIterator<Item = (&'a L, B)>,

Add edges to the lifetime graph.

This method is decoupled from LifetimeEnv::extend_lifetimes because generics can define new lifetimes, while where clauses cannot.

Panics

This method panics if any of the lifetime bounds arenโt already defined in the graph. This isnโt allowed by rustc in the first place, so it should only ever occur when deserializing an invalid LifetimeEnv.