diplomat_core::ast::lifetimes

Struct LifetimeTransitivity

pub struct LifetimeTransitivity<'env> {
    env: &'env LifetimeEnv,
    visited: Vec<bool>,
    out: Vec<&'env NamedLifetime>,
    longer_or_shorter: LongerOrShorter,
}

Collect all lifetimes that are either longer_or_shorter

Fields
env: &'env LifetimeEnvvisited: Vec<bool>out: Vec<&'env NamedLifetime>longer_or_shorter: LongerOrShorter
Implementations

impl<'env> LifetimeTransitivity<'env>

pub fn longer(env: &'env LifetimeEnv) -> Self

Returns a new LifetimeTransitivity that finds all longer lifetimes.

pub fn shorter(env: &'env LifetimeEnv) -> Self

Returns a new LifetimeTransitivity that finds all shorter lifetimes.

pub fn longer_than( env: &'env LifetimeEnv, named: &NamedLifetime, ) -> Vec<&'env NamedLifetime>

Returns all the lifetimes longer than a provided NamedLifetime.

pub fn shorter_than( env: &'env LifetimeEnv, named: &NamedLifetime, ) -> Vec<&'env NamedLifetime>

Returns all the lifetimes shorter than the provided NamedLifetime.

fn new(env: &'env LifetimeEnv, longer_or_shorter: LongerOrShorter) -> Self

Returns a new LifetimeTransitivity.

pub fn visit(&mut self, named: &NamedLifetime)

Visits a lifetime, as well as all the nodes it’s transitively longer or shorter than, depending on how the LifetimeTransitivity was constructed.

fn dfs(&mut self, index: usize)

Performs depth-first search through the LifetimeEnv created at construction for all nodes longer or shorter than the node at the provided index, depending on how the LifetimeTransitivity was constructed.

pub fn finish(self) -> Vec<&'env NamedLifetime>

Returns the transitively reachable lifetimes.