diplomat_core::ast

Module macros

Macros

accepted_tokens 🔒

define_macro_fragments 🔒

Structs

ItemList 🔒

MacroDef

Struct for defining a macro (i.e., macro_rules! example)

MacroIdent

Represents $Identifier:MacroFragSpec (see https://doc.rust-lang.org/reference/macros-by-example.html#railroad-MacroMatch)

MacroMatcher

Represents any given macro definition’s arguments, and information on how to parse them for use. A MacroMatcher is a delimited list of MacroMatches. When you call the macro example!(...), the macro matcher is the parentheses delimited tokens: (...). Used to compare a MacroDef against a MacroUse when parsing arguments.

MacroUse

A site where a macro is used (i.e., example!(...)) Constructed in MacroUse::parse (called by MacroDef::evaluate) when a previously defined macro is used. We use the definition of MacroDef to then construct the args in the used macro.

Macros

ParsePartial 🔒

Hack to read information from the TokenTree while saving the rest to be later copied into a buffer. Will attempt to read T, and on a success all of the unread tokens after the stream will be written to remaining.

Enums

MacroFrag

MacroMatch

A token representing part of a macro’s arguments. Used for determining how to parse MacroUse. First constructed with MacroMatcher::parse inside of MacroDef::parse. Then, we compare with MacroUse::parse_macro_matcher.