Module diplomat_core::ast::methods

Structs

BorrowedParams

Parameters in a method that might be borrowed in the return type.

Method

A method declared in the impl associated with an FFI struct. Includes both static and non-static methods, which can be distinguished by inspecting Method::self_param.

Param

A parameter taken by a Method, not including self.

SelfParam

The self parameter taken by a Method.

Enums

LifetimeKind

The type of lifetime.