diplomat_core::ast::methods

Struct Method

#[non_exhaustive]
pub struct Method {
    pub name: Ident,
    pub docs: Docs,
    pub full_path_name: Ident,
    pub self_param: Option<SelfParam>,
    pub params: Vec<Param>,
    pub return_type: Option<TypeName>,
    pub lifetime_env: LifetimeEnv,
    pub attrs: Attrs,
}

A method declared in the impl associated with an FFI struct. Includes both static and non-static methods, which can be distinguished by inspecting Method::self_param.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Ident

The name of the method as initially declared.

docs: Docs

Lines of documentation for the method.

full_path_name: Ident

The name of the FFI function wrapping around the method.

self_param: Option<SelfParam>

The self param of the method, if any.

params: Vec<Param>

All non-self params taken by the method.

return_type: Option<TypeName>

The return type of the method, if any.

lifetime_env: LifetimeEnv

The lifetimes introduced in this method and surrounding impl block.

attrs: Attrs

The list of cfg attributes (if any).

These are strings instead of syn::Attribute or proc_macro2::TokenStream because those types are not PartialEq, Hash, Serialize, etc.

Implementations

impl Method

pub fn from_syn( m: &ImplItemFn, self_path_type: PathType, impl_generics: Option<&Generics>, impl_attrs: &Attrs, ) -> Method

Extracts a Method from an AST node inside an impl.

pub fn borrowed_params(&self) -> BorrowedParams<'_>

Returns the parameters that the output is lifetime-bound to.

Examples

Given the following method:

fn foo<'a, 'b: 'a, 'c>(&'a self, bar: Bar<'b>, baz: Baz<'c>) -> FooBar<'a> { ... }

Then this method would return the &'a self and bar: Bar<'b> params because 'a is in the return type, and 'b must live at least as long as 'a. It wouldnโt include baz: Baz<'c> though, because the return type isnโt bound by 'c in any way.

Panics

This method may panic if TypeName::check_result_type_validity (called by Method::check_validity) doesnโt pass first, since the result type may contain elided lifetimes that we depend on for this method. The validity checks ensure that the return type doesnโt elide any lifetimes, ensuring that this method will produce correct results.

pub fn is_writeable_out(&self) -> bool

Checks whether the method qualifies for special writeable handling. To qualify, a method must:

not return any value

have the last argument be an &mut diplomat_runtime::DiplomatWriteable

Typically, methods of this form will be transformed in the bindings to a method that doesnโt take the writeable as an argument but instead creates one locally and just returns the final string.

pub fn has_writeable_param(&self) -> bool

Checks if any parameters are writeable (regardless of other compatibilities for writeable output)

pub fn docs(&self) -> &Docs

Returns the documentation block