diplomat_core::ast

Module modules

Structs

File

Module

ModuleBuilder 🔒

Contains all items needed to build an AST representation of a given Module, as we traverse through syn::ItemMod. We build this up in ModuleBuilder::add

Enums

DiplomatStructAttribute 🔒

Custom Diplomat attribute that can be placed on a struct definition.

DiplomatTypeAttribute 🔒

Custom Diplomat attribute that can be placed on an enum or struct definition.

Functions

extract_imports 🔒