diplomat_core::ast::modules

Struct ModuleBuilder

struct ModuleBuilder {
    custom_types_by_name: BTreeMap<Ident, CustomType>,
    custom_traits_by_name: BTreeMap<Ident, Trait>,
    functions_by_name: BTreeMap<Ident, Function>,
    sub_modules: Vec<Module>,
    imports: Vec<(Path, Ident)>,
    analyze_types: bool,
    type_parent_attrs: Attrs,
    impl_parent_attrs: Attrs,
    mod_macros: Macros,
}

Contains all items needed to build an AST representation of a given Module, as we traverse through syn::ItemMod. We build this up in ModuleBuilder::add

Fields
custom_types_by_name: BTreeMap<Ident, CustomType>custom_traits_by_name: BTreeMap<Ident, Trait>functions_by_name: BTreeMap<Ident, Function>sub_modules: Vec<Module>imports: Vec<(Path, Ident)>analyze_types: bool

As we traverse through the module, are we inside of #[diplomat::bridge]? If so, then analyze_types is set to true, and types, functions, and traits are all updated according to information parsed.

Otherwise, we traverse through modules until we find a module marked by #[diplomat::bridge]

type_parent_attrs: Attrsimpl_parent_attrs: Attrsmod_macros: Macros
Implementations

impl ModuleBuilder

fn add(&mut self, a: &Item)