diplomat_core::ast

Module types

Structs

LifetimeGenericsListDisplay 🔒

An fmt::Display type for formatting Rust lifetimes as they show up in generics list, when the generics list has no other elements

LifetimeGenericsListPartialDisplay 🔒

An fmt::Display type for formatting Rust lifetimes as they show up in generics list, when the generics list has another element

PathType

A named type that is just a path, e.g. std::borrow::Cow<'a, T>.

ReferenceDisplay 🔒

An fmt::Display type for formatting Rust references.

Enums

CustomType

A type declared inside a Diplomat-annotated module.

LifetimeOrigin

ModSymbol

A symbol declared in a module, which can either be a pointer to another path, or a custom type defined directly inside that module

Mutability

PrimitiveType

A built-in Rust primitive scalar type.

StdlibOrDiplomat

For types like Result/DiplomatResult, &[T]/DiplomatSlice<T> which can be specified using (non-ffi-safe) Rust stdlib types, or FFI-safe repr(C) types from diplomat_runtime, this tracks which of the two were used.

StringEncoding

TypeName

A local type reference, such as the type of a field, parameter, or return value. Unlike CustomType, which represents a type declaration, TypeNames can compose types through references and boxing, and can also capture unresolved paths.

Functions

get_lifetime_from_syn_path 🔒

get_ty_from_syn_path 🔒

is_runtime_type 🔒