diplomat_core::ast::types

Struct PathType

#[non_exhaustive]
pub struct PathType {
    pub path: Path,
    pub lifetimes: Vec<Lifetime>,
}

A named type that is just a path, e.g. std::borrow::Cow<'a, T>.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

path: Pathlifetimes: Vec<Lifetime>
Implementations

impl PathType

pub fn to_syn(&self) -> TypePath

pub fn new(path: Path) -> Self

pub fn extract_self_type(strct: &ItemStruct) -> Self

Get the Self type from a struct declaration.

Consider the following struct declaration:

struct RefList<'a> {
    data: &'a i32,
    next: Option<Box<Self>>,
}

When determining what type Self is in the next field, we would have to call this method on the syn::ItemStruct that represents this struct declaration. This method would then return a PathType representing RefList<'a>, so we know that’s what Self should refer to.

The reason this function exists though is so when we convert the fields’ types to PathTypes, we don’t panic. We don’t actually need to write the struct’s field types expanded in the macro, so this function is more for correctness,

pub fn resolve_with_path<'a>( &self, in_path: &Path, env: &'a Env, ) -> (Path, &'a CustomType)

If this is a TypeName::Named, grab the CustomType it points to from the env, which contains all CustomTypes across all FFI modules.

Also returns the path the CustomType is in (useful for resolving fields)

pub fn resolve<'a>(&self, in_path: &Path, env: &'a Env) -> &'a CustomType

If this is a TypeName::Named, grab the CustomType it points to from the env, which contains all CustomTypes across all FFI modules.

If you need to resolve struct fields later, call Self::resolve_with_path() instead to get the path to resolve the fields in.

pub fn trait_to_syn(&self) -> TraitBound

pub fn resolve_trait_with_path<'a>( &self, in_path: &Path, env: &'a Env, ) -> (Path, Trait)

pub fn resolve_trait<'a>(&self, in_path: &Path, env: &'a Env) -> Trait

If this is a TypeName::Named, grab the CustomType it points to from the env, which contains all CustomTypes across all FFI modules.

If you need to resolve struct fields later, call Self::resolve_with_path() instead to get the path to resolve the fields in.