diplomat_core::ast

Module types

Structs

PRIMITIVES_MAPPING 🔒

PRIMITIVE_TO_STRING 🔒

PathType

A named type that is just a path, e.g. std::borrow::Cow<'a, T>.

ReferenceDisplay 🔒

An fmt::Display type for formatting Rust references.

STRING_TO_PRIMITIVE 🔒

Enums

CustomType

A type declared inside a Diplomat-annotated module.

LifetimeOrigin

ModSymbol

A symbol declared in a module, which can either be a pointer to another path, or a custom type defined directly inside that module

Mutability

PrimitiveType

A built-in Rust primitive scalar type.

StringEncoding

TypeName

A local type reference, such as the type of a field, parameter, or return value. Unlike CustomType, which represents a type declaration, TypeNames can compose types through references and boxing, and can also capture unresolved paths.

Functions

is_runtime_type 🔒